Enable optional per-method Context callback.
Adds an optional callback to Service that allows manipulation of the
Context that is passed to the method handler based on the
service/method, as well as preliminary request rejection.

This avoids the need to place boilerplate code at the beginning of each
method handler and enables unified Service-wide security and setup
operations.